+++ HEP-data-entry/src/components/hepatitis/EntryField.tsx
@@ -1,5 +1,6 @@
 import { createElement } from "typed-html";
-import { SectionDataElement, CategoryOptionCombo, HelpTexts } from "../models/Form";
+import { HelpTexts } from "./models/Form";
+import { SectionDataElement, CategoryOptionCombo } from "../../models/d2Models";
 
 interface EntryFieldAttributes {
     dataElement: SectionDataElement;
