+++ HEP-data-entry/src/components/hepatitis/Table.tsx
@@ -1,14 +1,14 @@
 import { createElement } from "typed-html";
 import * as _ from "lodash";
 
+import { Form } from "./models/Form";
+import { EntryField } from "./EntryField";
 import {
     Section,
-    SectionDataElement,
-    Form,
     CategoryOptionCombo,
+    SectionDataElement,
     FieldsToRemove,
-} from "../models/Form";
-import { EntryField } from "./EntryField";
+} from "../../models/d2Models";
 
 interface TableAttributes {
     section: Section;
