+++ HEP-data-entry/src/components/hepatitis/models/Form.ts
@@ -1,54 +1,8 @@
-import { DataSet, Ref } from "../../../Dhis2Metadata";
-import { CustomForm } from "../components/CustomForm";
+import { CustomForm } from "../CustomForm";
 import * as _ from "lodash";
+import { Section, SectionDataElement, CategoryOptionCombo, OrderedSection } from "../../../models/d2Models";
+import { DataSet } from "../../../models/Dhis2Metadata";
 
-export interface DataEntryForm {
-    id: string;
-    name: string;
-    htmlCode: string;
-    style: "NORMAL" | "COMFORTABLE" | "COMPACT" | "NONE";
-}
-
-export interface Section {
-    id: string;
-    description: string;
-    displayName: string;
-    dataElements: SectionDataElement[];
-    formFields: OrderedSection;
-    greyedFields: GreyedFields[];
-    fieldsToRemove: FieldsToRemove[];
-}
-
-interface GreyedFields {
-    id: string;
-    dataElement: Ref;
-    categoryOptionCombo: Ref;
-}
-
-export interface FieldsToRemove {
-    dataElement: string;
-    categoryOptionCombo: string;
-}
-
-export interface SectionDataElement {
-    id: string;
-    code: string;
-    description?: string;
-    formName: string;
-    categoryCombo: { id: string; categoryOptionCombos: CategoryOptionCombo[] };
-    valueType: string;
-}
-
-export interface OrderedSection {
-    checkboxes: SectionDataElement[];
-    fields: SectionDataElement[];
-    globalEntry: SectionDataElement[];
-}
-
-export interface CategoryOptionCombo {
-    id: string;
-    name: string;
-}
 
 export interface HelpTexts {
     [key: string]: string;
